fdo#83057 File corrupts on save, as SDT is added incorrectly
A fly frame was attached to a para which started within a hint (run) containing an SDT.
This SDT was handled while exporting the FLYFRAME and also the text of the run.
So, eventhough the original file had only one sdt in the header, the RT file had two;
one for a seperate run as expected, and one incorrectly exported in the alternateContent (FLYFRAME)

So don't collect SDTPr from grabbag if the FLY is not processed.
As, the SDT will be handled when the run is exported.

Change-Id: I7b3c94208c171afbec54467fd6b756a6e30c816b
Reviewed-on: https://gerrit.libreoffice.org/11161
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Miklos Vajna <vmiklos@collabora.co.uk>