fdo#32613 RTF export: implemented anchored picture export
RTF originally didn't support anchored pictures: they were always
inline. Then it was invented that pictures can be exported anchored, if
they are in fact shapes, but their "pib" property is set to the hexdump
of the picture contents.

The RTF importer handled this situation for quite some time, but not the
exporter -- this commit implements that.

Change-Id: I173ad9f4f2e24620508ca63c0b892b7d5da4e5b2