n#708518: sd: check that master page matches when setting parent style
In ODF import it happened that the parent style of "outline2" etc.
was always set to the "outline1" style of the first master page in
the document, but it should be the "outline1" style of the same master
page as the "outline2".

(regression from 8f999ca4e55d892649694ce75cc9c6ace0a8cddf)

Change-Id: Ie563d5ee5c2040aeb6ca5c8bb25b195e15ea964e