Revert "force --with-system-odbc on MacOS X and Microsoft Windows"
There is no <sqlext.h> in the current OS X 10.9 SDK, or in the
upcoming 10.10 SDK.

Besides, if we want to enforce an option on a platform, we should just
make the configure script work as if the option had been present, not
require it to actually be passed to the configure script. The intent
is, to the best of my knowledge, that it should be possible to build
LibreOffice without passing any option at all to autogen.sh.

This reverts commit 367da5f82b0fa7ff62acf87db1becc63fef67449.

Change-Id: I0f72bc91bc51fa22ec994d7e39d2b10b9617834c