bnc#887307 ODF files not showing up on Windows search
Under Windows 7, ODF and OpenOffice.org 1.1 files were not found,
when the folder containing them was not indexed, and full text
search was switched on. The problem did not occur under Windows 8.

Apparently there is a bug in Windows 7, reading beyond the end of a
BufferStream resulted in 0 bytes read. So we take care not to read
beyond the end of stream.

Change-Id: Id01b08922121dabbb7b49c54f8ef11cb0e4ac413