fdo#81237: 2D OpenGL charts have wrong size
So far the size unit was 100th mm \ OPENGL_SCALE_VALUE
which is an absolute value, independent from the actual
size of the OpenGL window.
To solve that I added the SetSizePixel() method with
which we can specify the size in pixels and which
method resize the whole scene accordingly.

Change-Id: Ibd19a1bd9484c7bc8f8ce0d8113f60b95619db39