Resolves: fdo#80468 and fdo#81223 image/frame backgrounds wrong
writer has code to fill in missing new properties from the old properties
if they exist. But xmloff is stripping them out before they get there.

Don't strip them out, and add in a missing check for one of the
archaic bg colors and add a regression test for fdo#81223

Change-Id: I9a541a9bee0a01c90f2c33383f1144ecd8b0bfff