DOCX drawingML import: handle char color from theme for shape text
When we import table styles, we apply that as direct formatting, in case
there is no real direct formatting, see lcl_ApplyCellProperties() in the
sw UNO implementation.

We can do the same here: in case there is no other formatting, then
apply the char color from the WPS theme, that will give us the expected
result.

Change-Id: Ic8e6afc09167f7924a11ab0b445351075f16738e