proper scoping for docx redline import (bnc#821804)
propagateCharacterPropertiesAsSet sends the properties only when ending
a text run (or maybe starting another one, I'm not quite sure), so it breaks
ordering by sending them later then expected (although it worked in many cases).
It's a question if propagateCharacterPropertiesAsSet is to be used by anything
actually, since it seems rather broken to use it in the ooxml frontend.
Using sendPropertiesWithId sends the properties properly at the right time,
as one would expect. I don't know why dmapper can't simply handle this on its
own, as I think it does handle entering and leaving other elements, but
spending more time on it with this overdesigned abomination, oh well.

Change-Id: Ie36c5f933ea3e6d789ea8f9e4ee3b60a5d1c920c