Resolves:fdo#58277 don't double delete postits
In redline mode, if deleting the anchor point for a postit causes another one
to get moved, i.e. deleting it and creating a replacement. Then update the
the queue of postits to remove the old candidate and add the new candidate.

This is complicated by the occasional anonymous SWFMTFLD_REMOVED event
which just states that "something" got removed from the document (but not
truly deleted), and I don't see any better way to handle it than to completely
refill the pool of candidates and rely on the fix of

commit 41e83245167358cadd2f414b376bc3fbafe492ac
Date:   Thu Oct 2 16:33:46 2014 +0100
    Related: fdo#58277 only DelRight after a successful GotoField

to skip the previously successful candidates which got moved
into the redline backing area and so not loop forever

Change-Id: I0b30111bc1f2527011e68d048ecd65fcf71416a5