DOCX export: fix missing table border when original doc has table style
Scenario is that the original document has a table style, which sets
borders around the table, so A1 cell gets top and left borders (and none
for right/bottom). Then on export, if we write A1 as table default, then
this overwrites the table style, so the real right / bottom borders from
the table style are lost, as on export we think that they are available
from the table style already, so we don't write them.

Regression from commit b84ee2d7096d83422cf3958350be04b5d8e59154 (oox: Do
not overwrite table style properties, 2014-04-03).

Change-Id: I54c10ffc2ade65260693b769274b1311efc086f9