More hacking on 'icontest'
What we want is to test is not loading and displaying many images. It is just
to test just one. So simplify. (Yes, the name 'icontest' is even more
misleading now.)

Check whether the OpenGL context supports non-power-of-two textures, and also
check the maximum texture size, and scale / pad appropriately.

Change-Id: I02bccd33e08749d972652173a881aa40870c12d8