bnc#800714 SwTxtFrm::FormatAdjust: still call SplitFrm if we have columns
Regression from 26f2b4f3483250a9b0484f84f5a63cec2c284a89 (sw34bf05:
i#84870 - method <SwTxtFrm::FormatAdjust(..)> - do not split text frame
which only contains on as-character anchored object, 2011-03-11), the
i#84870 fix was for a document which didn't have columns, and didn't
consider that this causes a layout loop in case:

1) There are two paragraphs in a section, having multiple columns.

2) The second paragraph has fo:keep-together="always".

In this case originally we tried to call SplitFrm(), realized that it
didn't help, and then gave up. But after the change, we kept trying to
satisfy the two conflicting requirements (balance content in the 3
columns vs keep the second paragraph with the first one).

Fix the problem by not calling SplitFrm() only in case we're not inside
columns.

Change-Id: I64e969ef5e8f519314f5613f8e6fae626ae085ce