fdo#74441 [Accessiblity] Don't expose 0 for empty cells in Calc
At the moment non numercial cells in Calc is exposed to AT tools
(screen readers etc) as the value 0. It makes more sence to present
the string value of the cell which basically is what a sighted user
will see. This small patch should be a good start into making
Calc more accessible on Windows.

Change-Id: Ide76f2dc044c98f32aa78836049f12f2251522eb
Reviewed-on: https://gerrit.libreoffice.org/11996
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>