DOCX export: fix handling of paragraph mark on empty paragraphs
The problem was that in case a paragraph is empty, and it has both a
character style and direct formatting, then the doc model created by the
ODT and DOCX import is like this:

<text>
  <attrset>
    <item whichId="8" which="character font size" value="nHeight: 240, nProp: 100"/>
  </attrset>
  <hints>
    <hint end="0" whichId="52" which="character format" value="name: hello"/>
  </hints>
</text>

I.e. the direct formatting is stored in the attribute set of the text
node directly, while the character style is stored as a hint.

MSWordExportBase::OutputTextNode() tried to read the hints and if there
were not any, then tried to read the attribute set, while in this
situation we need both.

Change-Id: I71fd10f6d00246348e77fee5431cb4e2a7c9b349