Revert "Fix fdo#84795 Menu, DropDown-List not disappears with right mouse click"
This reverts commit e21e2ace3c8f9098e09abc09351c3ce1cee5a80f.

because it means that all disabled buttons get their "click" delivered
so disabled buttons are only visually disabled.

e.g. writer: insert->table->autoformat, click on the disabled "rename" or "add"
or "remove" and the sub dialog appears. Or in calc insert->object->chart and
click on the disabled "back" -> crash