Log the state as an integer, not as a byte with value 0, 1 or 2
It is rather hard to see in a terminal window whether a char with value 0, 1
or 2 has been printed;)

The elegant thing to do would of course be to make the enum in question into
an enum class, and define an operator<< for it, to output it symbolically. But
I can not be distracted into that now.

Change-Id: Ibfa4f4d5925d1db3cbbf35f84df70640f47a8d56